Setup CI pipeline
We use travis for our CI pipeline.

A helper script is used to setup the CI pipeline. The script creates
certificates, sets all required variables in the .env file and
configures the test environment.

The travis file first starts the application in daemon mode and then
runs the tests against the application. The build passes, if all tests
passed.